+++ projects/ids-enterprise-ng/src/lib/message/soho-message.d.ts
@@ -28,8 +28,14 @@ interface SohoMessageOptions {
   /** Width in pixels or auto. */
   width?: number | 'auto';
 
+  /** The type of icon to show. */
+  status?: string;
+
   /** Additional dialog styling. */
   cssClass?: string;
+
+  /** Set allowed tags */
+  allowedTags?: string;
 
   /** The buttons to create. */
   buttons?: SohoModalButton[];